Finished adding ScanObject ScriptableObjects
This system consists of two ScriptableObjects, the ScanItem and the ScanItemHolder. The ScanItemHolder is used for creating and storing the ScanItems, while ScanItems store data such as item names and descriptions.
ScanItems are stored in the ScanItemHolder so that they can be easily referenced via index - when telling the server that an item has been scanned/collected, you only have to send over an index rather than an entire ScriptableObject.